feat: add paper navigation buttons
