+++ src/lib/components/composites/paper-components/paper-view/cards/PaperDetailsCard.svelte
