+++ src/lib/components/composites/paper-components/paper-view/cards/PaperResearchContextCard.svelte
