+++ src/routes/paper/[paperId]/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import PaperView from "$lib/components/composites/paper-components/PaperView.svelte";
+    import PaperView from "$lib/components/composites/paper-components/paper-view/PaperView.svelte";
 
     const { data } = $props();
     const { user, paper } = data;
@@ -8,4 +8,4 @@
 <svelte:head>
     <title>{paper.title}</title>
 </svelte:head>
-<PaperView {user} {paper} />
+<PaperView {user} {paper} showNavigationButtons={false} />
